+++ eras/conway/impl/src/Cardano/Ledger/Conway/BlockBody.hs
@@ -4,9 +4,12 @@
 module Cardano.Ledger.Conway.BlockBody where
 
 import Cardano.Ledger.Alonzo.BlockBody
+import Cardano.Ledger.BaseTypes (ProtVer (..))
+import Cardano.Ledger.Binary (EncCBORGroup (..), serialize')
 import Cardano.Ledger.Conway.Era
 import Cardano.Ledger.Conway.Tx ()
 import Cardano.Ledger.Core
+import qualified Data.ByteString as BS
 
 instance EraBlockBody ConwayEra where
   type BlockBody ConwayEra = AlonzoBlockBody ConwayEra
@@ -14,3 +17,4 @@ instance EraBlockBody ConwayEra where
   txSeqBlockBodyL = txSeqBlockBodyAlonzoL
   hashBlockBody = alonzoBlockBodyHash
   numSegComponents = 4
+  blockBodySize (ProtVer v _) = BS.length . serialize' v . encCBORGroup