+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -11,6 +11,8 @@
 * Remove `ToCBOR` and `FromCBOR` instances for `PlutusWithContext`
   - Add `EncCBOR` and `DecCBOR` instances as replacement
 * Remove `ToCBOR (TxOut era)` superclass constraint from `EraTxOut`
+* Remove `EncCBORGroup (BlockBody era)` superclass constraint from `EraBlockBody`
+* Remove `bBodySize` in favour of the new `blockBodySize` method on `EraBlockBody`
 * Remove `validMetadatum`
 * Deprecate `BHeaderView` in favour `*EraBlockHeader` typeclasses.
   - Move `isOverlaySlot` to `Cardano.Ledger.Slot`.
@@ -36,6 +38,7 @@
 * Add `toPackageGolden` and `itGolden`
 * Add `goldenForToJSON`, `itGoldenToJSON`, `aesonGoldenSpec` and `roundTripAesonProperty`
 * Add `fullAnnCddlSpec`, `fullAnnGenCddlSpec`, `fullCddlSpec` and `fullGenCddlSpec`
+* Generalize the type of `genArrayTerm`
 * Add `TestBlockHeader` and `mkTestBlockHeaderNoNonce` as a replacement to deprecated `BHeaderView` and `makeHeaderView`.
 * Modify `ToExpr` instance for `Mismatch` to display type-level `r` parameter using `Typeable`